+++ Programming/Java/PalindromeOrNot/PalindromeCheck.java
@@ -1,0 +1,20 @@
+package com.raven7;
+
+import java.util.Scanner;
+
+class PalindromeCheck{
+    public static void main(String args[])
+    {
+        String original, reverse = ""; // Objects of String class
+        Scanner in = new Scanner(System.in);
+        System.out.println("Enter the number: ");
+        original = in.nextLine();
+        int length = original.length();
+        for ( int i = length - 1; i >= 0; i-- )
+            reverse = reverse + original.charAt(i);
+        if (original.equals(reverse))
+            System.out.println("True");
+        else
+            System.out.println("False");
+    }
+}
